agents/crypto-cex-bear/requirements.txt
# AI Model APIs
anthropic==0.25.0

# Binance Official API
binance-connector==3.6.0

# Telegram Bot
python-telegram-bot==21.0.0

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0

# Database (built-in, listed for reference)
# sqlite3 is included in Python stdlib
